requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
django-mysql==2.2.0
django-material==1.6.3
django-pwa==1.0.8
django-sslserver==0.22
django-progressive-web-app
